move _to_snake_case function to utils.py (#75)
