+++ supadantic/models.py
@@ -1,5 +1,4 @@
 import ast
-import re
 from abc import ABC
 from copy import copy
 from typing import TYPE_CHECKING, Any, ClassVar, Optional, TypeVar
@@ -10,6 +9,7 @@ from pydantic._internal._model_construction import ModelMetaclass as PydanticMod
 from .clients import SupabaseClient
 from .q_set import QSet
 from .query_builder import QueryBuilder
+from .utils import _to_snake_case
 
 
 if TYPE_CHECKING:
@@ -17,27 +17,6 @@ if TYPE_CHECKING:
 
 
 _M = TypeVar('_M', bound='BaseSBModel')
-
-
-def _to_snake_case(value: str) -> str:
-    """
-    Converts a string from camel case or Pascal case to snake case.
-
-    This function uses a regular expression to find uppercase letters within
-    the string and inserts an underscore before them (except at the beginning
-    of the string).  The entire string is then converted to lowercase.
-
-    Args:
-        value (str): The string to convert.
-
-    Returns:
-        (str): The snake_case version of the input string.
-
-    Example:
-        >>> _to_snake_case("MyClassName")
-        'my_class_name'
-    """
-    return re.sub(r'(?<!^)(?=[A-Z])', '_', value).lower()
 
 
 class ModelOptions:
